% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw_silverman.R
\name{bw_silverman}
\alias{bw_silverman}
\title{Calculate the bandwidth according to Silverman's rule of thumb.}
\usage{
bw_silverman(m)
}
\arguments{
\item{m}{A vector or a matrix. If a matrix is used as argument, Silverman's rule will be applied to each column separately.}
}
\value{
A vector of bandwidth values. If m is a vector, the function will return a single value. If m is a matrix of P columns, a vector of length P will be returned.
}
\description{
Silvermans rule of thumb uses the formula
bw = 0.9 * min(std.dev, IQR/1.34) * n^(-1/5)
}
\examples{
set.seed(123)
m <- matrix(c(3,3,10,2,6,5,4,6,9,10), ncol=2)
bw_silverman(m=m)
[1] 1.460377 1.688441
}
