% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dim_importance.R
\name{calculate_dim_importance}
\alias{calculate_dim_importance}
\title{Calculate the dimension importance score for each dimension}
\usage{
calculate_dim_importance(theta, party_pos)
}
\arguments{
\item{theta}{A vector with values of theta of length p}

\item{party_pos}{A data frame. Column 1 is the party name. I addition, one column should be included per dimension.}
}
\value{
A vector of dimension importance scores of length p
}
\description{
Given a model , the full data set used to train the model and a vector with the column positions of the already included variables, find the next best variable.
}
\examples{
party_pos <- data.frame(party=c("p1", "p2", "p3"), dim1=c(-2.5, 0, 0.1), dim2=c(0, -0.2, -1.7))
theta <- c(-1, -1)
calculate_dim_importance(theta=theta, party_pos=party_pos)
}
