% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_abilities_custom.R
\name{est_abilities_custom}
\alias{est_abilities_custom}
\title{Estimate the abilities given the item parameters and the test taker responses.}
\usage{
est_abilities_custom(
  item_params,
  newdata,
  the_method = "ML",
  norm_mean = NULL,
  norm_sd = NULL
)
}
\arguments{
\item{item_params}{A matrix of the item parameters. Each row represents one question item. The first column contains the discrimination parameter and the subsequent columns contain the difficulty parameters.}

\item{newdata}{A data frame or matrix with the response pattern to be evaluated. The number of items should equal the the number of the items in the matrix used as item_params.}

\item{norm_mean}{A vector of means for a custom normal prior. It should have the same length as there are observations in newdata.}

\item{norm_sd}{A vector of std devs for a custom normal prior. It should have the same length as there are observations in newdata.}

\item{method}{Can be 'ML' or 'MAP'. 'ML' is default.}
}
\value{
A vector of length N, where N is the number of test takers to evaluate.
}
\description{
Estimate the abilities given the item parameters and the test taker responses.
}
\examples{
Science <- mirt::Science
mirtmodel <- mirt(data=Science, model=1, itemtype = "graded")
my_item_params <- coef(mirtmodel, IRTpars=T, simplify=T)$items
est_abilities_custom(item_params=my_item_params, newdata=Science, the_method="ML")
}
