% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_abilities_mirt.R
\name{est_abilities_mirt}
\alias{est_abilities_mirt}
\title{Estimate the abilities given a model}
\usage{
est_abilities_mirt(model, newdata, the_method = "ML")
}
\arguments{
\item{model}{The mirt model.}

\item{newdata}{A data frame or matrix with the response pattern to be evaluated. The column names in newdata must be column names that are used in the dataset with which the MIRT model is trained.}

\item{the_metod}{"ML" or "MAP". "ML" is default.}
}
\value{
A vector of length N, where N is the number of test takers to evaluate.
}
\description{
Estimate the abilities given a MIRT model, using all or a subset of the items used to train the model..
}
\examples{
data(Science)
Science_model <- mirt(data = Science, model = 1, itemtype = "graded")
est_abilities_mirt(model = Science_model, newdata = Science[1:2, c(1, 3)])
}
