% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_next.R
\name{find_next}
\alias{find_next}
\title{Find the next best variable}
\usage{
find_next(model, model_data, included_variables = NULL)
}
\arguments{
\item{model}{The mirt model.}

\item{model_data}{A matrix or data frame with only one row. The column names should be the same as for the data training the model.}

\item{included_variables}{a vector with the positions of the columns of the included variables. If NULL (default), the current ability is estimated to 0.}
}
\value{
The function returns the name of the next best variable, the position of the column with the next best variable, the total information of the test with the next variable included and the information of each variable.
}
\description{
Given a model , the full data set used to train the model and a vector with the column positions of the already included variables, find the next best variable.
}
\examples{

Science_model <- mirt(data = Science, model = 1, itemtype = "graded")
find_next(model = Science_model, model_data=Science[1, , drop=F], included_variables = c(1,3))

}
