% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_next_custom.R
\name{find_next_custom}
\alias{find_next_custom}
\title{Find the next best variable with a set of item parameter}
\usage{
find_next_custom(theta, item_params, included_variables = NULL)
}
\arguments{
\item{included_variables}{a vector with the positions of the columns of the included variables. If NULL (default), the current ability is estimated to 0.}

\item{model}{The mirt model.}

\item{model_data}{A matrix or data frame with only one row. The column names should be the same as for the data training the model.}
}
\value{
The function returns the name of the next best variable, the position of the column with the next best variable, the total information of the test with the next variable included and the information of each variable.
}
\description{
Given a set of item parameter, the full data set used to train the model and a vector with the column positions of the already included variables, find the next best variable.
}
\examples{

Science <- mirt::Science
mirtmodel <- mirt(data=Science, model=1, itemtype = "graded")
my_item_params <- coef(mirtmodel, IRTpars=T, simplify=T)$items
theta <- 0
find_next_custom(theta, item_params=my_item_params)
find_next_custom(theta, item_params=my_item_params, included=3)
}
