% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_next_mv.R
\name{find_next_mv}
\alias{find_next_mv}
\title{Find the next best variable, choosing from variables in multiple dimensions.}
\usage{
find_next_mv(model, model_data, included_variables, party_pos)
}
\arguments{
\item{model}{a list of N mirt models.}

\item{model_data}{A list of N matrices or data frames, each with 1 row. The column names in the matrices/data frames should be the same as for the data training the model.}

\item{included_variables}{A list of N vectors, each with with the positions of the columns of the included variables in one of the dimensions. If NULL (default), the current ability is estimated to 0.}

\item{party_pos}{A data frame. Column 1 is the party name. I addition, one column should be included per dimension.}
}
\value{
The function returns the dimension with the next best variable, name of the next best variable, the position of the column with the next best variable, the total information of the test in each dimension with the next variable included. The information in One dimension will be updated, while the information in the other dimensions will remain unchanged.
}
\description{
Given N models, the full data set used to train each model and N vectors with the column positions of the already included variables, find the next best variable for each dimension.
}
\examples{

Science_model <- mirt(data = mirt::Science, model = 1, itemtype = "graded")
SAT12_model <- mirt(data = mirt::SAT12[, 1:5], model = 1, itemtype = "graded")
model <- list(Science_model, SAT12_model)
model_data <- list(Science[1, , drop=F], SAT12[1, 1:10, drop=F])
included_variables <- list(c(1, 3), NULL)
party_pos <- data.frame(party=c("p1", "p2", "p3"), dim1=c(-2.5, 0, 0.1), dim2=c(0, -0.2, -1.7))
find_next_mv(model = model, model_data=model_data, included_variables = included_variables)
}
