% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_next_mv_custom.R
\name{find_next_mv_custom}
\alias{find_next_mv_custom}
\title{Find the next best variable, choosing from variables in multiple dimensions.}
\usage{
find_next_mv_custom(theta, item_params, included_variables = NULL, party_pos)
}
\arguments{
\item{theta}{A vector of ability parameters in p dimensions.}

\item{item_params}{A list of item-parameter matrices. The list has length p.}

\item{included_variables}{A list of p vectors, one vector per dimension.}

\item{party_pos}{A data frame. Column 1 is the party name. I addition, one column should be included per dimension.}
}
\value{
The function returns the dimension with the next best variable, name of the next best variable, the position of the column with the next best variable, the total information of the test in each dimension with the next variable included. The information in One dimension will be updated, while the information in the other dimensions will remain unchanged.
}
\description{
Given N models, the full data set used to train each model and N vectors with the column positions of the already included variables, find the next best variable for each dimension.
}
\examples{

theta <- c(0,0)
Science_model <- mirt(data = mirt::Science, model = 1, itemtype = "graded")
SAT12_model <- mirt(data = mirt::SAT12[, 1:5], model = 1, itemtype = "graded")
items1 <- coef(Science_model, IRTpar=T, simplify=T)$items
items2 <- coef(SAT12_model, IRTpar=T, simplify=T)$items
item_params <- list(items1, items2)
included_variables <- list(c(1, 3), NULL)
party_pos <- data.frame(party=c("p1", "p2", "p3"), dim1=c(-2.5, 0, 0.1), dim2=c(0, -0.2, -1.7))
find_next_mv_custom(theta=theta, item_params=item_params, included_variables = included_variables, party_pos=party_pos)
}
