% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grm_info.R
\name{grm_info}
\alias{grm_info}
\title{Calculate the grm information}
\usage{
grm_info(theta, item_params, constant_D = F)
}
\arguments{
\item{theta}{The ability for which to estimate the information.}

\item{item_params}{A matrix of the item parameters. Each row represents one question item. The first column contains the discrimination parameter and the subsequent columns contain the difficulty parameters.}

\item{constant_D}{If true, include the constant D=1.702 in the formula for calculating the information. (Reckase 2009, page 52)}
}
\value{
A vector of length N, where N is the number of items included in item_params.
}
\description{
Calculate the information of one or multiple response items.
The unidimensional graded response model is assumed.
Note that the 2PL dichotomous model is a special case of the graded response model, with only two levels.
}
\examples{
my_item_params <- matrix(c(1, 1.5, -1, -0.5, 0, 1), ncol=3)
grm_info(theta=0, item_params=my_item_params)

}
