% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grm_info_multiple_items_abilities.R
\name{grm_info_multiple_items_abilities}
\alias{grm_info_multiple_items_abilities}
\title{Calculate the grm information}
\usage{
grm_info_multiple_items_abilities(theta, item_params, constant_D = F)
}
\arguments{
\item{theta}{The ability or abilities for which to estimate the information.}

\item{item_params}{matrix of the item parameters. The first column contains the discrimination parameters and the subsequent columns contain the difficulty parameters.}

\item{constant_D}{If true, include the constant D=1.702 in the formula for calculating the information. (Reckase 2009, page 52)}
}
\value{
A vector of length N, where N is the number of abilities included in theta.
}
\description{
In this version we apply the function to multiple ability levels at once.
}
\details{
Calculate the information of one or multiple response items.
The unidimensional graded response model is assumed.
Note that the 2PL dichotomous model is a special case of the graded response model, with only two levels.
}
\examples{
my_item_params <- matrix(c(1, 1.5, -1, -0.5, 0, 1), ncol=3)
grm_info_multiple_items_abilities(theta=c(-0.5, 0, 0.5, 1), item_params=my_item_params)

}
