% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dim_importance.R
\name{make_table_point_pairs}
\alias{make_table_point_pairs}
\title{Create a table with information about all pairs of party points}
\usage{
make_table_point_pairs(theta, points)
}
\arguments{
\item{theta}{A vector with values of theta}

\item{points}{A data frame which is the output of function make_table_points.}
}
\value{
a data frame with information for calculating dimension scores.
}
\description{
Currently, the function is written for 2 dimensions only.
}
\examples{
theta <- c(-1, 1)
party_pos <- data.frame(party=c("p1", "p2", "p3"), dim1=c(-2.5, 0, 0.1), dim2=c(0, -0.2, -1.7))
points <- make_table_points(theta=theta, party_pos=party_pos)
make_table_point_pairs(theta=theta, points=points)
}
