% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dim_importance.R
\name{make_table_points}
\alias{make_table_points}
\title{Create a table with information about the party points in each dimension}
\usage{
make_table_points(theta, party_pos)
}
\arguments{
\item{theta}{A vector with values of theta}

\item{party_pos}{A data frame. Column 1 is the party name. I addition, one column should be included per dimension.}
}
\value{
a data frame with columns for (1) party names (2) party positions in all dimensions (3) party distance to theta for each dimension (4) The euclidean distrances from each party to theta (5) for each dimension the change of the euclidean distance to theta as a function of a change in the dimension at the current location of theta.
}
\description{
Currently, the function is written for 2 dimensions only.
}
\examples{
theta <- c(-1, 1)
party_pos <- data.frame(party=c("p1", "p2", "p3"), dim1=c(-2.5, 0, 0.1), dim2=c(0, -0.2, -1.7))
make_table_points(theta=theta, party_pos=party_pos)
}
