% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_responses_grm.R
\name{simulate_response_grm}
\alias{simulate_response_grm}
\title{Simulate response data}
\usage{
simulate_response_grm(theta, item_params)
}
\arguments{
\item{theta}{A vector of theta values, representing the abilities of the simulated test takers.}

\item{item_params}{A matrix of the item parameters. Each row represents one question item. The first column contains the discrimination parameter and the subsequent columns contain the difficulty parameters.}
}
\value{
An A x B dimensioned data frame, where A is the number of simulated test takers, and B the number of question items. The values of the responses are in the interval [1, #levels], where #levels is determined by the argument item_params.
}
\description{
Simulate response data, given a vector of theta values and a matrix of item parameters.
}
\examples{
my_abilities <- rnorm(n=100)
my_item_params <- matrix(c(1, 1, 1, -1, 0, 1), ncol=2)
simulate_response_grm(theta=my_abilities, item_params=my_item_params)
}
